WIP - use c2rust to transpile libprio to Rust.
This needs a fair bit of manual cleanup and refactoring, but it compiles.

However, linking fails because libprio uses some internal code from
NSS, which was being compiled along with libprio before.

A few options for dealing with this that spring to mind, in order of
preference:

1. compile the bundled `mpi/` directory as before but compile into a separate
object then link with it
2. hack local copy of NSS so that it exposes the symbols we need
3. use the Rust conversion of `mpi/`, `src/mpi`

The last one is tempting but I am wary of diverging from NSS, the in-tree
copy is already difficult to maintain.